JetStream source refactoring
